Audio now works on Windows.
