+++ audio-backends/null/nullbackend.h
@@ -1,13 +1,13 @@
 #ifdef NULL_AUDIO
 
-#include "../../include/libmbGB/libmbgb.h"
 #include "../../include/libmbGB/apu.h"
+#include "../../include/libmbGB/libmbgb_api.h"
 using namespace gb;
 
 namespace gb
 {
-    void initaudio();
-    void deinitaudio();
+    void LIBMBGB_API initaudio();
+    void LIBMBGB_API deinitaudio();
 }
 
 #endif // NULL_AUDIO
