+++ audio-backends/sdl2/sdl2backend.h
@@ -1,14 +1,14 @@
 #ifdef SDL2_AUDIO
 
-#include "../../include/libmbGB/libmbgb.h"
 #include "../../include/libmbGB/apu.h"
+#include "../../include/libmbGB/libmbgb_api.h"
 #include <SDL2/SDL.h>
 using namespace gb;
 
 namespace gb
 {
-    void initsdl2audio();
-    void deinitsdl2audio();
+    void LIBMBGB_API initsdl2audio();
+    void LIBMBGB_API deinitsdl2audio();
 }
 
 #endif // SDL2_AUDIO
