+++ include/libmbGB/libmbgb_api.h
@@ -9,11 +9,9 @@
     #endif // LIBMBGB_EXPORTS
 
     #define LIBMBGB __cdecl
-    #define WIN32_AUDIO
 #else
     #define LIBMBGB_API
     #define LIBMBGB
-    #define UNIX_AUDIO
 #endif // _WIN32
 
 #endif // LIBMBGB_API_H
